requirements.txt
opencv-python
numpy
face-recognition
imutils
paho-mqtt
